Link hub to files
This is the trickiest part. Each hub in hubhub has a set of images and a set of documents that are references to some file in airtable. The gatsby-source-filesystem plugin creates its own quasi-standard definition for a `File` type in a gatsby schema, so we'll convert those links to the standard format using `createRemoteFileNode`. After creating the files we need to associate them with the correct field on the `Hub`, which is done in gatsby by appending `___NODE` to the end of the field name.