Add schema definition for a Hub
In order to make things in the schema match the expectations of our API exactly, we can make a custom type definition. Notably, certain fields are marked as non-null, and the file links are always available as well so they are marked as non-null arrays of non-null. Lastly, when making a custom type gatsby's defaults get overridden so you have to specify your linked nodes with `@link`.