Initial hubhub integration (#6)
* Add jsdoc types for API request

Add some jsdoc type info about the data we're working with so VSCode will help us out. If we want to switch to typescript at some point we can turn these into type definitions.

* sourceNodes scaffold

We're going to be using the sourceNodes hook to download data and put it into the site schema.

* Download hub data

To fill out our schema we'll fetch the hub data from the hubhub endpoint. We need to search for a specific hub name, which we will configure elsewhere. For now we'll just stick with something fake from the dev dataset that is deployed at this endpoint.

* Create a node with toplevel data

Pull off some fields that look relevant to this project and create a gatsby node. At this point you should be able to see the hub with [this localhost query](http://localhost:8000/__graphql?query=query%20MyQuery%20%7B%0A%20%20hub%20%7B%0A%20%20%20%20name%0A%20%20%7D%0A%7D%0A).

* Link hub to files

This is the trickiest part. Each hub in hubhub has a set of images and a set of documents that are references to some file in airtable. The gatsby-source-filesystem plugin creates its own quasi-standard definition for a `File` type in a gatsby schema, so we'll convert those links to the standard format using `createRemoteFileNode`. After creating the files we need to associate them with the correct field on the `Hub`, which is done in gatsby by appending `___NODE` to the end of the field name.

* Add schema definition for a Hub

In order to make things in the schema match the expectations of our API exactly, we can make a custom type definition. Notably, certain fields are marked as non-null, and the file links are always available as well so they are marked as non-null arrays of non-null. Lastly, when making a custom type gatsby's defaults get overridden so you have to specify your linked nodes with `@link`.

* fix typo