Add hero section (#4)
* Update various configs

1. Adding the styled-components gatsby plugin. This will precompile all the styles we create with styled-components. This makes the bundle smaller and avoids parsing the styles at runtime.
2. Adjusting the prettier config because closing brackets in JSX on a newline kind of bug me. Feel free to change this back if you don't agree it's not that big of a deal.

* Add photo placeholder

This is just a standard placeholder image. It can be nice to have while developing to get a sense for how something looks without having to search for a real photo.

* Remove width constraints on body so that hero can fill entire width

The top section of the main page needs to fill the entire width. We'll add the width constraint to a container further down in the tree.

* Implement initial hero section

This is the dense version of the design, and uses placeholders. Parameterization to come.

* Add dense state

In the figma the size of the hero depends on the presence of other content on the page. We can express this functionally and pass in a bool to toggle the density of the hero.

* Parameterize hero content

We know this data is eventually going to be queried from a source, so it makes sense to parameterize it now. Right now the image paths are strings but we can refactor them later to be gatsby image data.

* Make website optional

Not every hub will have their own website so this link should be optional.

* switch from css prop to styled tag