Appease the typechecker
This commit is just initial tidying. It's tough, because the JSDoc syntax can really be cumbersome at times but it's such a nice thing to have in place when you're dealing with data that comes in from a graphql query in gatsby and trying to map it onto your components.