Add about section (#7)
* Appease the typechecker

This commit is just initial tidying. It's tough, because the JSDoc syntax can really be cumbersome at times but it's such a nice thing to have in place when you're dealing with data that comes in from a graphql query in gatsby and trying to map it onto your components.

* Add About component

Just a basic component with the correct typography and spacing. We'll get fancier in a later commit.

* Add About to index

Places the about section on the page, and adds in a container to enforce the max width constraint from the design.

* Add newline parsing

This adds a markdown-style newline treatment where groups of 2 or more newline characters indicate a new paragraph. Also removes the default margin below the last paragraph so that outer containers can be responsible for layout-oriented spacing.