Initial hubhub wiring (#8)
* fields referring to files may not be present

* Look up hub by ID

Some hubs might have the same name, and also hub names are editable

* Load hero image data from hubhub

* Drop sunrise or sunrise movement from hub names

So we can stylize the name properly depending on the UI context

* Write graphql query for Hero

* Accept graphql data in Hero component

* Pass data in to hero from index

* Make the logo image fit a little nicer

`(Math.sqrt(2) / 2) * 100` sets the image to a square and fits its corners to the width of the circle. The padding on the container brings it in a bit. Setting object-fit to contain makes keeps the original proportions of the image.

* Load defaults from file system

These fields are optional in hubhub and won't be filled in for most hubs at least at first, so we need to have fallbacks.

* Wire up about section

* Made a separate test hub for us to use