One page for each hub (#10)
* Create hub nodes for all hubs in the data set

* Create a slug from the hub name

We need the slug to set up the URL for the page. Andrew is going to add an editable slug name to airtable but in the mean time we can just kebab-case the name.

* Create page for each hub

We're reusing the index page component for now. We'll switch over to a template later.

* move hub page template out of pages directory

* Add redirect from root

Since we're testing on netlify we can use netlify's _redirects feature to do this. We'll have to accomodate any other platform in a different way.